In order to properly remove a socket from a pollset, you must also provide the passed in reqevents that were given to pollset_add().  (ugh)
Additionally, kevent on Mac OS X will return a socket multiple times with the
same retevents flags if there is a read and write pair in the queue.  So, we
have to add code to track what events we've already seen for a socket in a
poll() run to avoid calling read() or write() multiple times when the socket
isn't ready.

* context.c
  (serf_connection_t): Add reqevents and seen_in_pollset fields.
  (update_pollset): Restore the reqevents from the previous invocation and
  save it before exiting.
  (open_connection): Clear out seen_in_pollset fields when we walk the
  connection lists.
  (serf_context_run): Check the seen_in_pollset bitmask to ensure we're
  not invoking a connection multiple times in the same pollset run.
  (remove_connection): Pass along reqevents to pollset_remove().
  (serf_connection_reset): If we reset a connection manually, ensure we
  don't see the APR_POLLHUP flag again for this socket.


git-svn-id: https://svn.apache.org/repos/asf/serf@1698552 13f79535-47bb-0310-9956-ffa450edef68