Start setting up a new build system based on SCons. Add a new SConstruct file.
Remove svn:ignore values for the autotools crap. Adjust the README to talk
about scons.

This builds a workable libserf-2.a that is usable by svn. The .dylib
version has some referential issues (rpath) to solve. And with the .a, we
need to communicate the proper libs to build with (not done). But the .a
library *does* work if compensations are made. IOW, this construct is good
ennough to build a functioning libserf.

* trunk/
* trunk/auth/
* trunk/buckets/
  (svn:ignore): remove the autoconf crap and ignore the new scons targets

* README:
  (...): talk about scons

* SConstruct:
  (...): new file


git-svn-id: https://svn.apache.org/repos/asf/serf@1699088 13f79535-47bb-0310-9956-ffa450edef68