Reorganize the chain construction a bit to avoid a cast.
Before applying the patch in r1589, I tweaked the code to always fetch
the chain (to avoid a compiler warning about uninitialized 'chain').
However, that causes a leak when building a one-cert chain. So... free
the old chain. (the original was fine in setting chain, but the
compiler didn't know that)

* buckets/ssl_buckets.c:
  (validate_server_certificate): free any possible fetched chain.
    reorganize the chain construction a bit. toss the later cast.


git-svn-id: https://svn.apache.org/repos/asf/serf@1699127 13f79535-47bb-0310-9956-ffa450edef68