Add a test to simulate an endless loop issue in serf_linebuf_fetch.
The specific case is a CRLF split state where the streams returns APR_EAGAIN
with 0 bytes read. serf_linebuf_fetch ignores this APR_EAGAIN and keeps peeking
the stream, which it shouldn't do according to the API rules.

The endless loop is easy to encounter in the test suite, where the stream can
only get new data when serf exits from serf_context_run so that the server
context can be triggered.

Note: this commit only adds the test and associated framework, it does not solve
the issue.

* buckets/buckets.c
  (serf_linebuf_fetch): Add note about the issue.

* test/test_serf.h
  (*): Add mock bucket declarations. Move some decls from test_buckets.c to
       here to make them available to other tests.

* test/test_buckets.c
  (CR, LF, CRLF, 
   read_and_check_bucket,
   readlines_and_check_bucket): Declare in test/test_serf.h.
  (readlines_and_check_bucket): Keep reading in case of SERF_NEWLINE_CRLF_SPLIT.
  (test_linebuf_crlf_split): New test.
  (test_buckets): Run the new test.

* test/mock_buckets.c: New bucket type, simulates source buckets (like socket).

* SConstruct
  (test_all): Add mock_buckets.c to compile target.

* test/test_all.c
  (testlist): Add commented out mock bucket test suite.


git-svn-id: https://svn.apache.org/repos/asf/serf@1699351 13f79535-47bb-0310-9956-ffa450edef68