Fix issue 114.
Scenario: the server triggers a ssl/tls renegotiation after receiving a request
but before writing a response.

Issue: The renegotiation request arrives in the ssl layer, and while reading
that data OpenSSL will write a response.
If the request was the only request pending on the connection (say one opening 
non-pipelined request), serf will have no more requests to send, and won't check
for socket writability. So the data that OpenSSL has add to the outgoing stream
will never be written on the socket, and after a timeout the server will abort
the connection.

The fix check for socket writability not only if there are requests pending,
but also if the outgoing stream has data pending.

* outgoing.c
  (request_or_data_pending): New function, checks if there's a request or if
       the outgoing stream contains data to send.
  (serf__conn_update_pollset): Use request_or_data_pending to decide when to
       set POLLOUT.
  (write_to_connection): Use request_or_data_pending to decide whether to enter
       the writing loop. Adapt some checks to handle the situation where there's
       data to send but no request pending. Connection is dirty when while
       reading the response of the last request, data was written on the
       outgoing stream.


git-svn-id: https://svn.apache.org/repos/asf/serf@1699615 13f79535-47bb-0310-9956-ffa450edef68