Simplify the ssl bucket code by removing a now unneeded layer of buffering and by reducing the BIO retry update/check code to just what is needed by OpenSSL.
Somehow our code assumed that the BIO retry flags we retrieved contained more
information than we set into it, while we were the only ones who update these
flags. The information we really needed is provided by the SSL layer on top
of the BIO layer, which might conflict with what is in our BIO implementation.
Let's use a simple boolean, as that is all we need.

Before this patch we first read all socket data through an intermediate stream
before handing it to OpenSSL, while the socket stream before us already buffers
everything efficiently enough for us.

* buckets/ssl_buckets.c
  (documentation): Update schema.
  (serf_ssl_stream_t): Remove variables that only apply to encode or decode.
  (serf_ssl_context_t): Add an encrypt/decrypt shared crypt_status as openssl
    assumes that we handle errors like some global 'errno' in the BIO layer.
    Add want_read boolean.

  (apps_ssl_info_callback): Update status usage.

  (bio_bucket_read): Unconditionally clear the BIO flags on entrance. Directly
    read from the decrypt stream as that is already buffered. Set retry flag
    when got an incomplete write for a different reason than EOF.

  (bio_bucket_write): Unconditionally clear the BIO flags and crypt status.

  (ssl_decrypt): Remove intermediate buffering. Always handle errors from
    SSL_read.

  (ssl_encrypt): Remove bio flag magic. Use the much simpler want_read boolean
    for stopping writes.

  (ssl_init_context): Update initialization.
  (serf_ssl_encrypt_destroy_and_data): Update cleanup.


git-svn-id: https://svn.apache.org/repos/asf/serf@1699766 13f79535-47bb-0310-9956-ffa450edef68