Fix the double linking problem in our test programs by explicitly linking against the static library.
The standard library search by name automatically enables linking
against both versions and the simple fix is to just pass the
library as a sourcefile.
(This solution was so simple that it took ages to find out)

This fix resolves several test failures on FreeBSD related to
having the same symbols in multiple locations.

* SConstruct
  (LIBNAME): Rename to...
  (SHLIBNAME): ... this and construct using platform defaults for shared
    libraries.
  (LIBNAMESTATIC): Rename to...
  (LIBNAME): ... this and construct using platform defaults for static
    libraries.

  Update several usages.

  (tenv): Remove LIBS reference to serf.
          Pass shared library with extension as additional source for tests.


git-svn-id: https://svn.apache.org/repos/asf/serf@1711833 13f79535-47bb-0310-9956-ffa450edef68