Add a few states to the deflate bucket state machine to allow using it for compression too.
Zlib already abstracts everything to a stream and we already had all the
necessary plumbing in place.

I think the original idea was to expose this via the same function as
deflating, but as we call this a deflate bucket while we only supported
inflating it is kind of hard to invent good names for the 'format' argument
to support compression.

* buckets/deflate_buckets.c
  (deflate_state_t): Add some initial and a finish state. Document memLevel usage.
  (serf_bucket_deflate_create): Tweak memlevel.
  (serf_bucket_deflate_compress_create): New function.
  (serf_deflate_refill): Tell ZLib that we have the final part. Only push back
    when necessary. Produce verify header when done.
  (serf_deflate_wait_for_data): Implement compress states.

* serf_bucket_types.h
  (serf_bucket_deflate_compress_create): New function.

* test/test_buckets.c
  (test_linebuf_fetch_crlf): Remove unused var.
  (test_deflate_compress_buckets): Add initial test.


git-svn-id: https://svn.apache.org/repos/asf/serf@1713889 13f79535-47bb-0310-9956-ffa450edef68