Make a standard 'bool' type available in our private API. This makes it easier to make some variables self-documenting.
We already use a similar construct in the mock http server, so this just makes
it easier to use a bool in our private functions without inventing our own
boolean type.

I don't see how we can expose this in the public api in a fully compatible
way, but this at least allows using it internally.

* serf_private.h
  (bool,
   true,
   false): Define if not already defined.


git-svn-id: https://svn.apache.org/repos/asf/serf@1714206 13f79535-47bb-0310-9956-ffa450edef68