On the '1.3.x-fix-outgoing-request-err' branch: Unset the hit_eof flag before reading from the pending output buckets.
With the current approach where a hold_open aggregate bucket is used to
check for the end of the outgoing data, the "hit_eof" flag can also be
inadvertently set when peeking at this bucket request_or_data_pending().
That would cause the actual read error (read_status) to be silently
ignored if the flag is has been set before the call to read_iovec().

Note: in trunk, the new serf_pump__write() function does the same thing.

* outgoing.c
  (write_to_connection): Unset the hit_eof flag before calling read_iovec().


git-svn-id: https://svn.apache.org/repos/asf/serf@1804543 13f79535-47bb-0310-9956-ffa450edef68